Changed subscription deletion handling.
Previously, when a subscription was deleted, I just removed the subscription from the collection. In the new implementation, when the deletion request event comes in, we first flag the subscription for deletion, and we move all non-billed consumption **immediately** into the to-be-reported collection. Once these values are successfully submitted, we remove the subscription from the collection.